+++ src/@types/i18next.d.ts
@@ -1,18 +1,19 @@
 import "i18next";
-import overview from "@public/locales/en/overview.json";
-import scan from "@public/locales/en/scan.json";
-import translation from "@public/locales/en/translation.json";
-import table from "@public/locales/en/table.json";
-import quarantine from "@public/locales/en/quarantine.json";
-import history from "@public/locales/en/history.json";
-import scheduler from "@public/locales/en/scheduler.json";
-import stats from "@public/locales/en/stats.json";
-import settings from "@public/locales/en/settings.json"
-import scanSettings from "@public/locales/en/scan-settings.json"
-import update from "@public/locales/en/update.json"
-import about from "@public/locales/en/about.json"
-import credits from "@public/locales/en/credits.json"
-import messages from "@public/locales/en/messages.json"
+import overview from "@i18n/en/overview.json";
+import scan from "@i18n/en/scan.json";
+import translation from "@i18n/en/translation.json";
+import table from "@i18n/en/table.json";
+import quarantine from "@i18n/en/quarantine.json";
+import history from "@i18n/en/history.json";
+import scheduler from "@i18n/en/scheduler.json";
+import stats from "@i18n/en/stats.json";
+import settings from "@i18n/en/settings.json"
+import scanSettings from "@i18n/en/scan-settings.json"
+import update from "@i18n/en/update.json"
+import about from "@i18n/en/about.json"
+import credits from "@i18n/en/credits.json"
+import messages from "@i18n/en/messages.json"
+import noClamAvPage from "@i18n/en/no-clamav-page.json"
 
 import { LangCode } from "../i18n/config";
 
@@ -34,6 +35,7 @@ declare module "i18next" {
                about: typeof about;
                credits: typeof credits;
                messages: typeof messages;
+               "no-clamav-page": typeof noClamAvPage;
           };
      }
 }