+++ src/components/layout/steps/installation.tsx
@@ -3,15 +3,17 @@ import { ExternalLink } from "lucide-react";
 import { Tabs, TabsContent, TabsList, TabsTrigger } from "@/components/ui/tabs";
 import CommandSnippetBlock from "@/components/command-snippet";
 import { openUrl } from "@tauri-apps/plugin-opener";
+import { useTranslation } from "react-i18next";
 
 export default function Installation(){
+     const {t} = useTranslation("no-clamav-page")
      return (
           <div className="space-y-3 text-left self-start px-6">
-               <p>Make sure to Install ClamAV by following the installation guide here:</p>
+               <p>{t("installation.info")}</p>
                <Button variant="link" onClick={()=>openUrl("https://docs.clamav.net/manual/Installing.html")}>
                     <ExternalLink/> docs.clamav.net/manual/Installing.html
                </Button>
-               <h2 className="text-xl md:text-2xl lg:text-3xl font-semibold">Recommended Linux Commands</h2>
+               <h2 className="text-xl md:text-2xl lg:text-3xl font-semibold">{t("installation.linux-commands")}</h2>
                <Tabs defaultValue="deb">
                     <TabsList className="w-full">
                          <TabsTrigger value="rpm">Fedora, OpenSUSE, etc</TabsTrigger>