+++ CHANGELOG.md
@@ -18,6 +18,7 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
   - Credits
 - Toast messages and validation messages
 - Notifications
+- The **No ClamAV Found** Page
 ### Improved
 - Some more Enums used in frontend
 - `console.error()` in some places into `toast.error(...,{description})`